10/04/2018 Quadratic Function
Create a c program that evaluates a quadratic function x^2 +5x +6.  The main program will call a function passing the value for x (accepted as input) and the function returns the evaluated value. The evaluated value should be printed in the main program along with the x value. Output should be as follows, with at least the five different user inputs. The program should work repeatedly as long as the input is between -100.0 and 100.0 (end points not included).