10/11/2018 Standard Deviation
Develop a C program that will perform the following operations on an array containing grades by using appropriate functions. Use the array elements (as shown in the output) in the main program and call the respective functions. The main function should print the array values before the array change and array values after the array change.
The three different functions respectively should 
1. calculate the standard deviation and return the standard deviation.
2. add the standard deviation (use the integer portion after rounding) to each of the original array elements. If the new value exceeds 100, the final value should be limited to 100.
3. Sort the array using modified bubble sort. Instead of making the same number of comparisons for every pass, modify it so that for every successive pass, it makes one less comparison from the previous pass.